Edit for public use
The major structural change is to show the two parts of the tutorial
as one coherent whole. There is no need to consider only people
who attended the first part of the original tutorial any more.

A link to a sample solution to Part 1 has been provided. It is a file
on Drive right now.

I fixed a mistake in the suggestions for how to add time to the timer.